UserService-CreateContactMethod: allow using existing contact methods
Previously, if the new contact method is duplicate, the method will fail. This change allows to first
check if the existing contact method and the new one to be added is the same, if so, fetch the existing one
and return successfully.
This feature is useful in terraform provider, as it can sync existing contact methods automatically to terraform
states.