update replacing obsolete test case `TestUsersAddDuplicateContactMethod`
`TestUsersAddDuplicateContactMethod` is based on the premise that adding
duplicated contact methods is actually allowed and desired, being that
this is an unwanted behaviour and the error "User Contact method must be
unique" should be returned.

Introduces `TestUsersErrorWhenContactMethodExists` as test case to
validate the expected result in this case, which is returning the error
"User Contact method must be unique".